Overview/Hackathon2/oncoflow-ai-main/requirements.txt
streamlit
torch
transformers
sentence-transformers
faiss-cpu
pypdf
numpy
pandas
sacrebleu
rouge-score
spacy